dom/tests/mochitest/bugs/test_bug633133.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633133
-->
<head>
  <title>Test for Bug 633133</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633133">Mozilla Bug 633133</a>
<p id="display"></p>
<div id="content" style="display: none">
  <div id='foo'></div>
  <div name='bar'></div>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 633133 **/

var divCollection = document.getElementsByTagName('div');

ok("foo" in divCollection, "'foo' should be in the div collection");
ok("bar" in divCollection, "'bar' should be in the div collection");
ok(!("" in divCollection), "empty string shouldn't be in the div collection");
ok(!("foobar" in divCollection), "'foobar' shouldn't be in the div collection");

</script>
</pre>
</body>
</html>
